Move Blocks, NavBar, TopicImage to templates directory
