+++ src/templates/Blocks/StatBlock.jsx
