+++ src/templates/Footer/Branding.js
