+++ src/theme/styles/index.scss
@@ -1,4 +1,9 @@
+// Component styles.
 @import "../../components/FileDownload/filedownload.scss";
+@import "../../components/IconList/iconlist.scss";
 @import "../../components/IconListItem/iconlistitem.scss";
-@import "../../components/NavBar/navbar.scss";
 @import "../../components/Organization/organization.scss";
+
+// Template styles.
+@import "../../templates/Hero/hero.scss";
+@import "../../templates/NavBar/navbar.scss";
