+++ stories/search.stories.jsx
@@ -75,7 +75,7 @@ storiesOf('Search', module)
   .add('Search Results Message', () => (
     <SearchResultsMessage
       searchTerm={text('Search Term', '')}
-      total={number('Total Results', 10)}
+      total={number('Total Results', 1000)}
       facetLimit={number('Facet Limit', 3)}
       facetDelimiter={text('Facet Delimiter', ' or ')}
       facetSeparator={text('Facet Separator', ' | ')}
